updated web notification modal and button
